requirements.txt
flask==3.0.2
flask-cors==4.0.1
jinja2==3.1.4
requests==2.31.0
gevent==24.2.1
icecream
python-dotenv